+++ src/main/php/com/openai/rest/ApiEndpoint.class.php
@@ -1,28 +1,15 @@
 <?php namespace com\openai\rest;
 
-use webservices\rest\Endpoint;
+use lang\Value;
+use util\Comparison;
+use util\log\Traceable;
 
-/** Base class for AzureAI and OpenAI endpoints */
-abstract class ApiEndpoint {
-  protected $endpoint;
+/** Base class for Distributed, AzureAI and OpenAI endpoints */
+abstract class ApiEndpoint implements Traceable, Value {
+  use Comparison;
 
-  /** Creates a new API endpoint */
-  public function __construct(Endpoint $endpoint) { $this->endpoint= $endpoint; }
-
-  /** @return util.URI */
-  public function base() { return $this->endpoint->base(); }
-
-  /** @return [:var] */
-  public function headers() { return $this->endpoint->headers(); }
-
-  /**
-   * Provides a log category for tracing requests
-   *
-   * @param  ?util.log.LogCategory $cat
-   */
-  public function setTrace($cat) {
-    $this->endpoint->setTrace($cat);
-  }
+  /** Returns rate limit */
+  public abstract function rateLimit(): RateLimit;
 
   /** Returns an API */
   public abstract function api(string $path, array $segments= []): Api;