+++ src/main/php/com/openai/rest/AzureAIEndpoint.class.php
@@ -4,11 +4,14 @@ use util\URI;
 use webservices\rest\Endpoint;
 
 /**
- * Azure OpenAI REST API endpoint
+ * Azure OpenAI APIs. Usage slightly differs from their public OpenAI counterpart,
+ * as the API Key must be passed in a special header, and an API version must be
+ * included in the query string.
  *
  * @test com.openai.unittest.AzureAIEndpointTest
  */
 class AzureAIEndpoint extends ApiEndpoint {
+  private $endpoint, $rateLimit;
   public $version;
 
   /**
@@ -19,17 +22,39 @@ class AzureAIEndpoint extends ApiEndpoint {
    */
   public function __construct($arg, $version= null) {
     if ($arg instanceof Endpoint) {
-      parent::__construct($arg);
+      $this->endpoint= $arg;
       $this->version= $version;
     } else {
       $uri= $arg instanceof URI ? $arg : new URI($arg);
       $this->version= $version ?? $uri->param('api-version');
-      parent::__construct((new Endpoint($uri))->with(['Authorization' => null, 'API-Key' => $uri->user()]));
+      $this->endpoint= (new Endpoint($uri))->with(['Authorization' => null, 'API-Key' => $uri->user()]);
     }
+    $this->rateLimit= new RateLimit();
+  }
+
+  /** Returns rate limit */
+  public function rateLimit(): RateLimit { return $this->rateLimit; }
+
+  /** @return [:var] */
+  public function headers() { return $this->endpoint->headers(); }
+
+  /**
+   * Provides a log category for tracing requests
+   *
+   * @param  ?util.log.LogCategory $cat
+   */
+  public function setTrace($cat) {
+    $this->endpoint->setTrace($cat);
   }
 
   /** Returns an API */
   public function api(string $path, array $segments= []): Api {
-    return new Api($this->endpoint->resource(ltrim($path, '/').'?api-version='.$this->version, $segments));
+    return new Api(
+      $this->endpoint->resource(ltrim($path, '/').'?api-version='.$this->version, $segments),
+      $this->rateLimit
+    );
   }
+
+  /** @return string */
+  public function toString() { return nameof($this).'(->'.$this->endpoint->base().'?api-version='.$this->version.')'; }
 }