+++ src/main/php/com/openai/rest/OpenAIEndpoint.class.php
@@ -9,6 +9,7 @@ use webservices\rest\Endpoint;
  * @test com.openai.unittest.OpenAIEndpointTest
  */
 class OpenAIEndpoint extends ApiEndpoint {
+  private $endpoint, $rateLimit;
 
   /**
    * Creates a new OpenAI endpoint
@@ -18,7 +19,8 @@ class OpenAIEndpoint extends ApiEndpoint {
    * @param  ?string $project
    */
   public function __construct($arg, $organization= null, $project= null) {
-    parent::__construct($arg instanceof Endpoint ? $arg : new Endpoint($arg));
+    $this->endpoint= $arg instanceof Endpoint ? $arg : new Endpoint($arg);
+    $this->rateLimit= new RateLimit();
 
     // Pass optional organization and project IDs
     $headers= [];
@@ -27,8 +29,26 @@ class OpenAIEndpoint extends ApiEndpoint {
     $headers && $this->endpoint->with($headers);
   }
 
+  /** Returns rate limit */
+  public function rateLimit(): RateLimit { return $this->rateLimit; }
+
+  /** @return [:var] */
+  public function headers() { return $this->endpoint->headers(); }
+
+  /**
+   * Provides a log category for tracing requests
+   *
+   * @param  ?util.log.LogCategory $cat
+   */
+  public function setTrace($cat) {
+    $this->endpoint->setTrace($cat);
+  }
+
   /** Returns an API */
   public function api(string $path, array $segments= []): Api {
-    return new Api($this->endpoint->resource(ltrim($path, '/'), $segments));
+    return new Api($this->endpoint->resource(ltrim($path, '/'), $segments), $this->rateLimit);
   }
+
+  /** @return string */
+  public function toString() { return nameof($this).'(->'.$this->endpoint->base().')'; }
 }